Initial execution strategy work
