+++ src/main/java/graphql/execution/ExecutionStrategy.java
@@ -1,0 +1,142 @@
+package graphql.execution;
+
+import graphql.ExceptionWhileDataFetching;
+import graphql.ExecutionResult;
+import graphql.GraphQLException;
+import graphql.language.Field;
+import graphql.schema.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+import static graphql.introspection.Introspection.SchemaMetaFieldDef;
+import static graphql.introspection.Introspection.TypeMetaFieldDef;
+import static graphql.introspection.Introspection.TypeNameMetaFieldDef;
+
+public abstract class ExecutionStrategy {
+    private static final Logger log = LoggerFactory.getLogger(ExecutionStrategy.class);
+
+    protected ValuesResolver valuesResolver = new ValuesResolver();
+    protected FieldCollector fieldCollector = new FieldCollector();
+
+    abstract ExecutionResult execute(ExecutionContext executionContext, GraphQLObjectType parentType, Object source, Map<String, List<Field>> fields);
+
+    protected Object resolveField(ExecutionContext executionContext, GraphQLObjectType parentType, Object source, List<Field> fields) {
+        GraphQLFieldDefinition fieldDef = getFieldDef(executionContext.getGraphQLSchema(), parentType, fields.get(0));
+        if (fieldDef == null) return null;
+
+        Map<String, Object> argumentValues = valuesResolver.getArgumentValues(fieldDef.getArguments(), fields.get(0).getArguments(), executionContext.getVariables());
+        DataFetchingEnvironment environment = new DataFetchingEnvironment(
+                source,
+                argumentValues,
+                executionContext.getRoot(),
+                fields,
+                fieldDef.getType(),
+                parentType,
+                executionContext.getGraphQLSchema()
+        );
+
+        Object resolvedValue = null;
+        try {
+            resolvedValue = fieldDef.getDataFetcher().get(environment);
+        } catch (Exception e) {
+            log.info("Exception while fetching data", e);
+            executionContext.addError(new ExceptionWhileDataFetching(e));
+        }
+
+        return completeValue(executionContext, fieldDef.getType(), fields, resolvedValue);
+    }
+
+    protected Object completeValue(ExecutionContext executionContext, GraphQLType fieldType, List<Field> fields, Object result) {
+        if (fieldType instanceof GraphQLNonNull) {
+            GraphQLNonNull graphQLNonNull = (GraphQLNonNull) fieldType;
+            Object completed = completeValue(executionContext, graphQLNonNull.getWrappedType(), fields, result);
+            if (completed == null) throw new GraphQLException("Cannot return null for non-nullable type: " + fields);
+            return completed;
+
+        } else if (result == null) {
+            return null;
+        } else if (fieldType instanceof GraphQLList) {
+            return completeValueForList(executionContext, (GraphQLList) fieldType, fields, (List<Object>) result);
+        } else if (fieldType instanceof GraphQLScalarType) {
+            return completeValueForScalar((GraphQLScalarType) fieldType, result);
+        } else if (fieldType instanceof GraphQLEnumType) {
+            return completeValueForEnum((GraphQLEnumType) fieldType, result);
+        }
+
+
+        GraphQLObjectType resolvedType;
+        if (fieldType instanceof GraphQLInterfaceType) {
+            resolvedType = resolveType((GraphQLInterfaceType) fieldType, result);
+        } else if (fieldType instanceof GraphQLUnionType) {
+            resolvedType = resolveType((GraphQLUnionType) fieldType, result);
+        } else {
+            resolvedType = (GraphQLObjectType) fieldType;
+        }
+
+        Map<String, List<Field>> subFields = new LinkedHashMap<>();
+        List<String> visitedFragments = new ArrayList<>();
+        for (Field field : fields) {
+            if (field.getSelectionSet() == null) continue;
+            fieldCollector.collectFields(executionContext, resolvedType, field.getSelectionSet(), visitedFragments, subFields);
+        }
+
+        // Calling this from the executionContext so that you can shift from the simple execution strategy for mutations
+        // back to the desired strategy.
+
+        return executionContext.getExecutionStrategy().execute(executionContext, resolvedType, result, subFields).getData();
+    }
+
+    protected GraphQLObjectType resolveType(GraphQLInterfaceType graphQLInterfaceType, Object value) {
+        GraphQLObjectType result = graphQLInterfaceType.getTypeResolver().getType(value);
+        if (result == null) throw new GraphQLException("could not determine type");
+        return result;
+    }
+
+    protected GraphQLObjectType resolveType(GraphQLUnionType graphQLUnionType, Object value) {
+        GraphQLObjectType result = graphQLUnionType.getTypeResolver().getType(value);
+        if (result == null) throw new GraphQLException("could not determine type");
+        return result;
+    }
+
+
+    protected Object completeValueForEnum(GraphQLEnumType enumType, Object result) {
+        return enumType.getCoercing().coerce(result);
+    }
+
+    protected Object completeValueForScalar(GraphQLScalarType scalarType, Object result) {
+        return scalarType.getCoercing().coerce(result);
+    }
+
+    protected Object completeValueForList(ExecutionContext executionContext, GraphQLList fieldType, List<Field> fields, List<Object> result) {
+        List<Object> completedResults = new ArrayList<>();
+        for (Object item : result) {
+            completedResults.add(completeValue(executionContext, fieldType.getWrappedType(), fields, item));
+        }
+        return completedResults;
+    }
+
+    protected GraphQLFieldDefinition getFieldDef(GraphQLSchema schema, GraphQLObjectType parentType, Field field) {
+        if (schema.getQueryType() == parentType) {
+            if (field.getName().equals(SchemaMetaFieldDef.getName())) {
+                return SchemaMetaFieldDef;
+            }
+            if (field.getName().equals(TypeMetaFieldDef.getName())) {
+                return TypeMetaFieldDef;
+            }
+        }
+        if (field.getName().equals(TypeNameMetaFieldDef.getName())) {
+            return TypeNameMetaFieldDef;
+        }
+
+        GraphQLFieldDefinition fieldDefinition = parentType.getFieldDefinition(field.getName());
+        if (fieldDefinition == null) throw new GraphQLException("unknown field " + field.getName());
+        return fieldDefinition;
+    }
+
+
+}
