+++ src/main/java/graphql/execution/SimpleExecutionStrategy.java
@@ -1,0 +1,23 @@
+package graphql.execution;
+
+import graphql.ExecutionResult;
+import graphql.ExecutionResultImpl;
+import graphql.language.Field;
+import graphql.schema.GraphQLObjectType;
+
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+public class SimpleExecutionStrategy extends ExecutionStrategy {
+    @Override
+    public ExecutionResult execute(ExecutionContext executionContext, GraphQLObjectType parentType, Object source, Map<String, List<Field>> fields) {
+        Map<String, Object> results = new LinkedHashMap<>();
+        for (String fieldName : fields.keySet()) {
+            List<Field> fieldList = fields.get(fieldName);
+            Object resolvedResult = resolveField(executionContext, parentType, source, fieldList);
+            results.put(fieldName, resolvedResult);
+        }
+        return new ExecutionResultImpl(results, executionContext.getErrors());
+    }
+}
